feat: add regex support for banned words
