+++ packages/identifier/src/banned.ts
@@ -1,0 +1,98 @@
+/*
+ * This file is meant for 100% no-go, banned words that should have fuzzier matching in order to catch more usernames (ie,
+ * explicitly use regexes before and after each of these strings in order to catch all instances of these words being used.)
+ */
+
+// naive, incomplete list of slurs and unacceptable words
+const slurs = [
+  'bluegum',
+  'chink',
+  'chinks',
+  'coolie',
+  'coolies',
+  'coon',
+  'coons',
+  'coont',
+  'golliwog',
+  'golliwogs',
+  'gook',
+  'gooks',
+  'gyp',
+  'gyps',
+  'half-breed',
+  'halfbreed',
+  'half-breeds',
+  'halfbreeds',
+  'heeb',
+  'heebs',
+  'hitler',
+  'kaffer',
+  'kaffers',
+  'kaffir',
+  'kaffirs',
+  'kaffre',
+  'kaffres',
+  'kafir',
+  'kafirs',
+  'kike',
+  'kikes',
+  'kkk',
+  'klukluxklan',
+  'muzzie',
+  'n1gga',
+  'n1gger',
+  'nazi ',
+  'negorid',
+  'negress',
+  'nig',
+  'nigg3r',
+  'nigg4h',
+  'nigga',
+  'niggah',
+  'niggas',
+  'niggaz',
+  'nigger',
+  'niggerachi',
+  'niggerette',
+  'niggerican',
+  'niggerino',
+  'niggeroid',
+  'niggers',
+  'nigglet',
+  'nigguh',
+  'nigguhs',
+  'nig-nog',
+  'nig-nogs',
+  'nigs',
+  'paki',
+  'pakis',
+  'pedophile',
+  'pickaninnie',
+  'pickaninnies',
+  'pickaninny',
+  'pickaninnys',
+  'raghead',
+  'ragheads',
+  'redskin',
+  'sambo',
+  'sambos',
+  'spade',
+  'spades',
+  'spic',
+  'spics',
+  'squaw',
+  'squaws',
+  'wetback',
+  'wetbacks',
+  'yid',
+  'yids',
+]
+
+export const bannedSubdomains: Record<string, boolean> = [
+  ...slurs,
+].reduce((acc, cur) => {
+  return {
+    ...acc,
+    [cur]: true,
+  }
+}, {})
