+++ packages/identifier/src/reserved.ts
@@ -1047,96 +1047,10 @@ const famousAccounts = [
   'zerohora',
 ]
 
-// naive, incomplete list of slurs and unacceptable words
-const slurs = [
-  'bluegum',
-  'chink',
-  'chinks',
-  'coolie',
-  'coolies',
-  'coon',
-  'coons',
-  'coont',
-  'golliwog',
-  'golliwogs',
-  'gook',
-  'gooks',
-  'gyp',
-  'gyps',
-  'half-breed',
-  'halfbreed',
-  'half-breeds',
-  'halfbreeds',
-  'heeb',
-  'heebs',
-  'hitler',
-  'kaffer',
-  'kaffers',
-  'kaffir',
-  'kaffirs',
-  'kaffre',
-  'kaffres',
-  'kafir',
-  'kafirs',
-  'kike',
-  'kikes',
-  'kkk',
-  'klukluxklan',
-  'muzzie',
-  'n1gga',
-  'n1gger',
-  'nazi ',
-  'negorid',
-  'negress',
-  'nig',
-  'nigg3r',
-  'nigg4h',
-  'nigga',
-  'niggah',
-  'niggas',
-  'niggaz',
-  'nigger',
-  'niggerachi',
-  'niggerette',
-  'niggerican',
-  'niggerino',
-  'niggeroid',
-  'niggers',
-  'nigglet',
-  'nigguh',
-  'nigguhs',
-  'nig-nog',
-  'nig-nogs',
-  'nigs',
-  'paki',
-  'pakis',
-  'pedophile',
-  'pickaninnie',
-  'pickaninnies',
-  'pickaninny',
-  'pickaninnys',
-  'raghead',
-  'ragheads',
-  'redskin',
-  'sambo',
-  'sambos',
-  'spade',
-  'spades',
-  'spic',
-  'spics',
-  'squaw',
-  'squaws',
-  'wetback',
-  'wetbacks',
-  'yid',
-  'yids',
-]
-
 export const reservedSubdomains: Record<string, boolean> = [
   ...atpSpecific,
   ...commonlyReserved,
   ...famousAccounts,
-  ...slurs,
 ].reduce((acc, cur) => {
   return {
     ...acc,
